2200/15 - added version digital as v3 (#481)
* added version digital as v3

* added version digital as v3

* added version digital as v3

* temp

* added visit id col

* updated digitial regressions

* reran regressions

* visit date merged main
