+++ cishouseholds/derive.py
@@ -62,7 +62,7 @@ def count_barcode_cleaned(
     df: DataFrame, column_name_to_assign: str, barcode_column: str, date_taken_column: str, visit_datetime_colum: str
 ):
     """
-    Count occurances of barcode
+    Count occurrences of barcode
     Parameters
     ----------
     df
@@ -1184,7 +1184,7 @@ def assign_column_regex_match(
         Name of column that will be matched
     pattern
         Regular expression pattern as a string
-        Needs to be a raw string literal (preceeded by r"")
+        Needs to be a raw string literal (preceded by r"")
 
     Returns
     -------
