+++ cishouseholds/pipeline/input_variable_names.py
@@ -158,7 +158,7 @@ column_name_maps = {
         "Suffix": "suffix",
         "Full Name": "full_name",
         "DoB": "date_of_birth",
-        "Email": "email",
+        "Email": "email_address",
         "No Email Address": "no_email_address",
         "Bloods Taken": "bloods_taken",
         "Bloods Barcode 1": "blood_sample_barcode",
@@ -220,7 +220,7 @@ column_name_maps = {
         "Previous Symptoms-Diarrhoea": "symptoms_since_last_visit_diarrhoea",
         "Previous Symptoms-Loss of taste": "symptoms_since_last_visit_loss_of_taste",
         "Previous Symptoms-Loss of smell": "symptoms_since_last_visit_loss_of_smell",
-        "If Yes; Date of first symptoms": "think_had_covid_date",
+        "If Yes; Date of first symptoms": "think_had_covid_onset_date",
         "Did you contact NHS?": "think_had_covid_contacted_nhs",
         "If Yes; Were you tested": "other_pcr_test_since_last_visit",
         "If Yes;Test Result": "other_pcr_test_results",
@@ -284,7 +284,7 @@ column_name_maps = {
         "Middle_Name": "middle_name",
         "Last_Name": "last_name",
         "DoB": "date_of_birth",
-        "Email": "email",
+        "Email": "email_address",
         "No_Email_address": "no_email_address",
         "Bloods_Taken": "bloods_taken",
         "bloods_barcode_1": "blood_sample_barcode",
@@ -374,7 +374,7 @@ column_name_maps = {
         "Previous_Symptoms_Diarrhoea": "symptoms_since_last_visit_diarrhoea",
         "Previous_Symptoms_Loss_of_taste": "symptoms_since_last_visit_loss_of_taste",
         "Previous_Symptoms_Loss_of_smell": "symptoms_since_last_visit_loss_of_smell",
-        "If_yes_Date_of_first_symptoms": "think_had_covid_date",
+        "If_yes_Date_of_first_symptoms": "think_had_covid_onset_date",
         "Did_you_contact_NHS": "think_had_covid_contacted_nhs",
         "Were_you_admitted_to_hospital": "think_had_covid_admitted_to_hospital",
         "Have_you_had_a_swab_test": "other_pcr_test_since_last_visit",
@@ -456,7 +456,7 @@ column_name_maps = {
         "Middle_Name": "middle_name",
         "Last_Name": "last_name",
         "DoB": "date_of_birth",
-        "Email": "email",
+        "Email": "email_address",
         "Have_landline_number": "have_landline_number",
         "Have_mobile_number": "have_mobile_number",
         "Have_email_address": "have_email_address",
@@ -597,7 +597,7 @@ column_name_maps = {
         "Previous_Symptoms_Low_mood_not_enjoying_anything": "symptoms_since_last_visit_low_mood_or_not_enjoying_anything",  # noqa: E501
         "Previous_Symptoms_Memory_loss_or_confusion": "symptoms_since_last_visit_memory_loss_or_confusion",
         "Previous_Symptoms_Difficulty_concentrating": "symptoms_since_last_visit_difficulty_concentrating",
-        "If_yes_Date_of_first_symptoms": "think_had_covid_date",
+        "If_yes_Date_of_first_symptoms": "think_had_covid_onset_date",
         "Did_you_contact_NHS": "think_had_covid_contacted_nhs",
         "Were_you_admitted_to_hospital": "think_had_covid_admitted_to_hospital",
         "Have_you_had_a_swab_test": "other_pcr_test_since_last_visit",
