+++ cishouseholds/pipeline/timestamp_map.py
@@ -3,7 +3,7 @@ survey_responses_v2_datetime_map = {
         "symptoms_last_7_days_onset_date",
         "last_covid_contact_date",
         "last_suspected_covid_contact_date",
-        "think_had_covid_date",
+        "think_had_covid_onset_date",
         "other_pcr_test_first_positive_date",
         "other_pcr_test_last_negative_date",
         "other_antibody_test_first_positive_date",
@@ -25,7 +25,7 @@ survey_responses_v1_datetime_map = {
         "symptoms_last_7_days_onset_date",
         "last_covid_contact_date",
         "last_suspected_covid_contact_date",
-        "think_had_covid_date",
+        "think_had_covid_onset_date",
         "other_pcr_test_first_positive_date",
         "other_pcr_test_last_negative_date",
         "cis_covid_vaccine_date",
@@ -44,7 +44,7 @@ survey_responses_v0_datetime_map = {
         "symptoms_last_7_days_onset_date",
         "last_covid_contact_date",
         "last_suspected_covid_contact_date",
-        "think_had_covid_date",
+        "think_had_covid_onset_date",
         "date_of_birth",
     ],
     "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'": ["visit_datetime", "samples_taken_datetime"],
