+++ tests/pipeline/test_survey_responses_version_0/processed_responses_v0_schema_regression.yml
@@ -220,7 +220,7 @@ fields:
   nullable: true
   type: timestamp
 - metadata: {}
-  name: email
+  name: email_address
   nullable: true
   type: string
 - metadata: {}
@@ -464,7 +464,7 @@ fields:
   nullable: true
   type: string
 - metadata: {}
-  name: think_had_covid_date
+  name: think_had_covid_onset_date
   nullable: true
   type: timestamp
 - metadata: {}
@@ -543,10 +543,6 @@ fields:
   name: outward_postcode
   nullable: true
   type: string
-- metadata: {}
-  name: consent_summary
-  nullable: true
-  type: double
 - metadata: {}
   name: swab_taken
   nullable: false
