+++ tests/pipeline/test_survey_responses_version_1/processed_responses_v1_schema_regression.yml
@@ -224,7 +224,7 @@ fields:
   nullable: true
   type: timestamp
 - metadata: {}
-  name: email
+  name: email_address
   nullable: true
   type: string
 - metadata: {}
@@ -584,7 +584,7 @@ fields:
   nullable: true
   type: string
 - metadata: {}
-  name: think_had_covid_date
+  name: think_had_covid_onset_date
   nullable: true
   type: timestamp
 - metadata: {}
@@ -715,10 +715,6 @@ fields:
   name: outward_postcode
   nullable: true
   type: string
-- metadata: {}
-  name: consent_summary
-  nullable: true
-  type: double
 - metadata: {}
   name: swab_taken
   nullable: false
