+++ tests/pipeline/test_survey_responses_version_2/processed_responses_v2_schema_regression.yml
@@ -252,7 +252,7 @@ fields:
   nullable: true
   type: timestamp
 - metadata: {}
-  name: email
+  name: email_address
   nullable: true
   type: string
 - metadata: {}
@@ -816,7 +816,7 @@ fields:
   nullable: true
   type: string
 - metadata: {}
-  name: think_had_covid_date
+  name: think_had_covid_onset_date
   nullable: true
   type: timestamp
 - metadata: {}
@@ -1103,10 +1103,6 @@ fields:
   name: outward_postcode
   nullable: true
   type: string
-- metadata: {}
-  name: consent_summary
-  nullable: true
-  type: double
 - metadata: {}
   name: swab_taken
   nullable: false
